pyproject.toml
[project]
name = "codex-aura"
version = "0.1.0"
description = "Code context mapping for AI agents"
authors = [{name = "Lambertain", email = "your.email@example.com"}]
license = {text = "Apache-2.0"}
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pyyaml>=6.0",
    "pathspec>=0.12.0",
    "slowapi>=0.1.9",
    "structlog>=23.1.0",
    "prometheus-client>=0.19.0",
    "opentelemetry-distro>=0.45b0",
    "opentelemetry-instrumentation-fastapi>=0.45b0",
    "opentelemetry-exporter-jaeger>=1.21.0",
    "psutil>=5.9.0"
]

[project.scripts]
codex-aura = "codex_aura.cli.main:main"

[tool.poetry]
name = "codex-aura"
version = "0.1.0"
description = "Code context mapping for AI agents"
authors = ["Lambertain <your.email@example.com>"]
license = "Apache-2.0"
python = "^3.11"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
pydantic = "^2.5.0"
pyyaml = "^6.0"
pathspec = "^0.12.0"
structlog = "^23.1.0"
prometheus-client = "^0.19.0"
opentelemetry-distro = "^0.45b0"
opentelemetry-instrumentation-fastapi = "^0.45b0"
opentelemetry-exporter-jaeger = "^1.21.0"
psutil = "^5.9.0"

[tool.poetry.group.dev.dependencies]
pytest = "*"
pytest-cov = "*"
ruff = "*"
mypy = "*"

[tool.poetry.scripts]
codex-aura = "codex_aura.cli.main:main"

[project.entry-points."codex_aura.plugins.context"]
basic = "codex_aura.plugins.builtin.context_basic:BasicContextPlugin"

[project.entry-points."codex_aura.plugins.impact"]
basic = "codex_aura.plugins.builtin.impact_basic:BasicImpactPlugin"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "--cov=codex_aura --cov-report=term-missing --cov-report=html --cov-fail-under=80"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"